vendor: perf: Conditionally update to LA.VENDOR.1.0.r1
Blobs are updated to LA.VENDOR.1.0.r1-16000-WAIPIO.QSSI13.0
Configs are updated to to LA.VENDOR.1.0.r1-15600-WAIPIO.0

This does not work yet on pre-5.10 SoCs.
Those SoCs will use the perf-legacy
component until that is fixed (or indefinitely
depending on the complexity of the fix).

Note: I am aware that there are conditions in perf
that only apply to legacy and conditions in perf-legacy
that only apply to perf. We need to nail down the
supported targets before I can clean up completely.

Change-Id: I7925042da82fac6e5353c1a4b1660b3ca9e128d7